requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.10
django-filter==2.2.0
djangorestframework==3.10.3
filelock==3.0.12
importlib-metadata==1.5.0
importlib-resources==1.0.2
kombu==4.6.7
Markdown==3.2.1
pipenv==2018.11.26
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
vine==1.3.0
virtualenv==20.0.1
virtualenv-clone==0.5.3
zipp==2.2.0
